Refactor Order-related classes and interfaces for better
