+++ src/main/java/com/coralblocks/coralme/OrderBookListener.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright 2023 (c) CoralBlocks - http://www.coralblocks.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,24 +15,24 @@
  */
 package com.coralblocks.coralme;
 
-import com.coralblocks.coralme.Order.CancelReason;
-import com.coralblocks.coralme.Order.ExecuteSide;
-import com.coralblocks.coralme.Order.RejectReason;
+import com.coralblocks.coralme.CancelReason;
+import com.coralblocks.coralme.ExecuteSide;
+import com.coralblocks.coralme.RejectReason;
 
 public interface OrderBookListener {
-    
+
     public void onOrderReduced(OrderBook orderBook, long time, Order order, long reduceNewTotalSize);
-    
+
     public void onOrderCanceled(OrderBook orderBook, long time, Order order, CancelReason cancelReason);
-    
+
     public void onOrderExecuted(OrderBook orderBook, long time, Order order, ExecuteSide executeSide, long executeSize, long executePrice, long executeId, long executeMatchId);
-    
+
     public void onOrderAccepted(OrderBook orderBook, long time, Order order);
-    
+
     public void onOrderRejected(OrderBook orderBook, long time, Order order, RejectReason rejectReason);
-    
+
     public void onOrderRested(OrderBook orderBook, long time, Order order, long restSize, long restPrice);
-    
+
     public void onOrderTerminated(OrderBook orderBook, long time, Order order);
-    
+
 }
